Normalize: Use size_t where appropriate; timestamp is uint64_t
timestamp is uint64_t
- AttPDUMsg

Use size_t where appropriate
- OctetTypes.hpp: TROOctets*
- ATTPDUTypes.hpp: AttPDUMsg*
- HCITypes.hpp: HCIPacket*
- HCIComm.hpp: HCIComm
- L2CAPComm.hpp: L2CAPComm
- MgmtTypes.hpp: MgmtCommand*, MgmtEvent*
- UUID.hpp: uuid_t
- DBTManager, GATTHandler, HCIHandler

- HCIComm.hpp: HCIComm
- BTTypes.hpp
- DBTTypes.hpp

+++ api/direct_bt/BTTypes.hpp
@@ -444,7 +444,7 @@ namespace direct_bt {
         POctets data;
 
         ManufactureSpecificData(uint16_t const company) noexcept;
-        ManufactureSpecificData(uint16_t const company, uint8_t const * const data, int const data_len) noexcept;
+        ManufactureSpecificData(uint16_t const company, uint8_t const * const data, size_t const data_len) noexcept;
 
         std::string toString() const noexcept;
     };
@@ -578,7 +578,7 @@ namespace direct_bt {
          * https://www.bluetooth.com/specifications/archived-specifications/
          * </p>
          */
-        static std::vector<std::shared_ptr<EInfoReport>> read_ad_reports(uint8_t const * data, uint8_t const data_length) noexcept;
+        static std::vector<std::shared_ptr<EInfoReport>> read_ad_reports(uint8_t const * data, size_t const data_length) noexcept;
 
         /**
          * Reads the Extended Inquiry Response (EIR) or Advertising Data (AD) segments
